Add JSON-LD Local Business (#26)
This adds support for[ basic local business](https://developers.google.com/search/docs/data-types/local-business) properties as outlined in the readme. 

By default the type will be LocalBusiness but users can provide any valid sub-type as per the documentation. 

Images code was taken from product but not turned into a generic helper. This felt like it could be reused but leaving them separate allowed for better isolation. Happy to change this and move buildImages to a common builders file with tests. 

Additional changes to reorder imports/exports alphabetically and add the missing product link in the readme are included in the same commit.   

Fixes #23